Update README.md
The project is a data analysis of an e-commerce store using descriptive data analysis to describe how the store has performed over the year. Amongst the discoveries of the data analyst are
1. The Top 5 Suburb group.
2. Top 5 buyers each year.
3. Top Best-Selling Products.
4. The best-selling managers across the suburb.
5. Top 5 Best-Selling Days.
6. Top 5 Best Selling by Month.

The analysis showed that a particular suburb had the best sales over some time and that clients trooped in from a particular suburb to purchase their products.
These analyses will help the management of the e-commerce store to make decisive decisions on which products to increase in stock. It will also assist them in making a critical decision on opening another branch in the particular suburb in the area where they have the most patronage.